requirements.txt
pandas>=2.0.0
numpy>=1.24.0
sqlalchemy>=2.0.0
pyodbc>=4.0.39
python-dotenv>=1.0.0